src/test_func/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --extra test_func --no-hashes
aiolimiter==1.2.1
annotated-types==0.7.0
anyio==4.8.0
apscheduler==3.11.0
beautifulsoup4==4.9.3
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.4.1
decorator==5.1.1
exceptiongroup==1.2.2 ; python_full_version < '3.11'
google-api-core==2.24.0
google-auth==2.19.1
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.5.0
google-cloud-pubsub==2.16.0
google-cloud-secret-manager==2.16.0
googleapis-common-protos==1.66.0
greenlet==3.1.1 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.7
lxml==4.9.1
nest-asyncio==1.6.0
proto-plus==1.25.0
protobuf==4.25.5
psycopg2-binary==2.9.5
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.5
pydantic-core==2.27.2
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-telegram-bot==21.11.1
requests==2.32.3
retry2==0.9.5
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==1.4.54
typing-extensions==4.12.2
tzdata==2024.2 ; sys_platform == 'win32'
tzlocal==5.2
urllib3==1.26.19
